Create view mixin for tooltips
This particular mixin initializes the foundation tooltips. It is used
for the tooltips that are included in the template of the view, that
extends the mixin. It works and for elements that are rendered
dynamically. Lastly, the mixin removes all the tooltips of the view,
when the view is about to be destroyed.